Fix new branch cretion in empty repo
