+++ plugins/scaffolder-backend/src/scaffolder/actions/builtin/publish/index.ts
@@ -18,10 +18,6 @@ export { createPublishAzureAction } from './azure';
 export { createPublishBitbucketAction } from './bitbucket';
 export { createPublishBitbucketCloudAction } from './bitbucketCloud';
 export { createPublishBitbucketServerAction } from './bitbucketServer';
-export { createPublishBitbucketServerRepoCreateAction } from './bitbucketServerRepoCreate';
-export { createPublishBitbucketServerRepoCloneAction } from './bitbucketServerRepoClone';
-export { createPublishBitbucketServerPushAction } from './bitbucketServerPush';
-export { createPublishBitbucketServerPullRequestAction } from './bitbucketServerPullRequest';
 export { createPublishGerritAction } from './gerrit';
 export { createPublishGerritReviewAction } from './gerritReview';
 export { createPublishGithubAction } from './github';
